new RequestCard structure
