+++ src/components/requestCard/bottom/bottom.scss
@@ -1,0 +1,43 @@
+@import "shared-colors";
+
+.bottom {
+	background-color: $black;
+	border-radius: 0px 0px 5px 5px;
+	border: 1px solid $black;
+	border-top: none;
+	padding: 10px 15px;
+	color: aliceblue;
+	user-select: none;
+
+	display: flex;
+	justify-content: space-between;
+	align-items: center;
+	font-weight: bold;
+	cursor: pointer;
+	transition: all 0.2s ease-out;
+
+	&.blue {
+		background-color: $blue;
+		border: 1px solid $blue-dark-1;
+	}
+
+	&.green {
+		background-color: $green;
+		border: 1px solid $green-dark-1;
+	}
+
+	&.orange {
+		background-color: $orange;
+		border: 1px solid $orange-dark-1;
+	}
+
+	&.purple {
+		background-color: $purple;
+		border: 1px solid $purple-dark-1;
+	}
+
+	&.red {
+		background-color: $red;
+		border: 1px solid $red-dark-1;
+	}
+}
