+++ src/components/requestCard/content/content.scss
@@ -1,0 +1,12 @@
+@import "shared-colors";
+
+.content {
+	padding: 4px;
+	border-style: solid;
+	border-color: $lightgray;
+	border-width: 0px 1px;
+
+	.responseWrapper {
+		margin: 10px;
+	}
+}
