+++ src/components/requestCard/content/index.tsx
@@ -1,0 +1,71 @@
+import * as React from "react";
+import s from "./content.scss";
+import MonacoEditor from "react-monaco-editor";
+import { componentSwitch } from "helpers/componentSwitch";
+import ReactJson from "react-json-view";
+import { requestModel } from "helpers/requestModel";
+import { TabKeys } from "../tabs";
+
+interface ContentProps {
+	activeTab: TabKeys;
+	args: any;
+	setJsonArgs: React.Dispatch<any>;
+	model: requestModel;
+}
+
+export default function Content(props: ContentProps) {
+	const { activeTab, args, setJsonArgs, model } = props;
+
+	const Content = componentSwitch<TabKeys>(activeTab, {
+		arguments: (
+			<MonacoEditor
+				height="250"
+				language="json"
+				theme="vs-light"
+				defaultValue={JSON.stringify(args, null, 2)}
+				options={{
+					minimap: {
+						enabled: false,
+						showSlider: "mouseover",
+						renderCharacters: false,
+					},
+				}}
+				onChange={v => {
+					try {
+						setJsonArgs(JSON.parse(v));
+					} catch (err) {
+						console.log("JS-1", err);
+					}
+				}}
+			/>
+		),
+		response: (
+			<div className={s.responseWrapper}>
+				<ReactJson src={model.response !== undefined ? model.response : {}} name={false} />
+			</div>
+		),
+		extra: (
+			<div className={s.responseWrapper}>
+				<p>
+					Did you know this is a <b>work in progress?</b>
+				</p>
+				<p>... well, now you know.</p>
+			</div>
+		),
+		error: (
+			<div className={s.responseWrapper}>
+				{model.error ? <pre>{model.error}</pre> : <b>There is no errors</b>}
+			</div>
+		),
+		default: (
+			<div className={s.responseWrapper}>
+				<p>
+					Did you know this is a <b>work in progress?</b>
+				</p>
+				<p>... well, now you know.</p>
+			</div>
+		),
+	});
+
+	return <div className={s.content}>{Content}</div>;
+}
