+++ src/components/requestCard/header/header.scss
@@ -1,0 +1,66 @@
+@import "shared-colors";
+
+.callName {
+	display: flex;
+	align-items: center;
+
+	.hrefIcon {
+		margin-left: 10px;
+		color: $c3;
+		margin: 5px;
+		&:hover {
+			color: $c8;
+		}
+	}
+}
+
+.statusCircle {
+	margin-right: 10px;
+	transition: all 0.2s ease-out;
+
+	&.blue {
+		color: $blue;
+	}
+
+	&.green {
+		color: $green;
+	}
+
+	&.orange {
+		color: $orange;
+	}
+
+	&.purple {
+		color: $purple;
+	}
+
+	&.red {
+		color: $red;
+	}
+}
+
+.header {
+	border-radius: 5px 5px 0px 0px;
+	border: 1px solid $grey;
+	padding: 10px 15px;
+
+	display: flex;
+	justify-content: space-between;
+	align-items: center;
+	padding: 10px 15px;
+	cursor: pointer;
+
+	.callName {
+		display: flex;
+		align-items: center;
+
+		.hrefIcon {
+			margin-left: 10px;
+			color: $c3;
+			margin: 5px;
+			&:hover {
+				color: $c8;
+			}
+		}
+	}
+}
