+++ src/components/requestCard/requestCard.scss
@@ -4,101 +4,6 @@
 	max-width: 100%;
 	height: 100%;
 	margin-bottom: 20px;
-
-	.header {
-		border-radius: 5px 5px 0px 0px;
-		border: 1px solid $grey;
-		padding: 10px 15px;
-
-		display: flex;
-		justify-content: space-between;
-		align-items: center;
-		padding: 10px 15px;
-		cursor: pointer;
-
-		.callName {
-			display: flex;
-			align-items: center;
-
-			.hrefIcon {
-				margin-left: 10px;
-				color: $c3;
-				margin: 5px;
-				&:hover {
-					color: $c8;
-				}
-			}
-		}
-	}
-
-	.subHeader {
-		display: flex;
-		border: 1px solid $lightgray;
-		border-top: none;
-		.tab {
-			cursor: pointer;
-			padding: 10px 15px;
-			font-weight: bold;
-			font-size: 12px;
-			transform: translateY(1px);
-			color: $midgnightBlue;
-			&.active {
-				border-bottom: 2px solid $blue;
-				color: $blue;
-			}
-		}
-	}
-	.content {
-		padding: 4px;
-		border-style: solid;
-		border-color: $lightgray;
-		border-width: 0px 1px;
-
-		.responseWrapper {
-			margin: 10px;
-		}
-	}
-	.bottom {
-		background-color: $black;
-		border-radius: 0px 0px 5px 5px;
-		border: 1px solid $black;
-		border-top: none;
-		padding: 10px 15px;
-		color: aliceblue;
-		user-select: none;
-
-		display: flex;
-		justify-content: space-between;
-		align-items: center;
-		font-weight: bold;
-		cursor: pointer;
-		transition: all 0.2s ease-out;
-
-		&.blue {
-			background-color: $blue;
-			border: 1px solid $blue-dark-1;
-		}
-
-		&.green {
-			background-color: $green;
-			border: 1px solid $green-dark-1;
-		}
-
-		&.orange {
-			background-color: $orange;
-			border: 1px solid $orange-dark-1;
-		}
-
-		&.purple {
-			background-color: $purple;
-			border: 1px solid $purple-dark-1;
-		}
-
-		&.red {
-			background-color: $red;
-			border: 1px solid $red-dark-1;
-		}
-	}
 }
 .closedCard {
 	max-width: 100%;
@@ -111,45 +16,4 @@
 	align-items: center;
 	padding: 10px 15px;
 	cursor: pointer;
-	.callName {
-		display: flex;
-		align-items: center;
-
-		.hrefIcon {
-			margin-left: 10px;
-			color: $c3;
-			margin: 5px;
-			&:hover {
-				color: $c8;
-			}
-		}
-	}
-
-	.openIcon {
-	}
-
-	.statusCircle {
-		margin-right: 10px;
-		transition: all 0.2s ease-out;
-
-		&.blue {
-			color: $blue;
-		}
-
-		&.green {
-			color: $green;
-		}
-
-		&.orange {
-			color: $orange;
-		}
-
-		&.purple {
-			color: $purple;
-		}
-
-		&.red {
-			color: $red;
-		}
-	}
 }
