+++ src/components/requestCard/tabs/tabs.scss
@@ -1,0 +1,19 @@
+@import "shared-colors";
+
+.subHeader {
+	display: flex;
+	border: 1px solid $lightgray;
+	border-top: none;
+	.tab {
+		cursor: pointer;
+		padding: 10px 15px;
+		font-weight: bold;
+		font-size: 12px;
+		transform: translateY(1px);
+		color: $midgnightBlue;
+		&.active {
+			border-bottom: 2px solid $blue;
+			color: $blue;
+		}
+	}
+}
